fix compile

+++ app/src/main/java/app/gyrolet/mpvrx/ui/celestial/CelestialEffects.kt
@@ -85,7 +85,7 @@ fun Modifier.celestialBorder(shape: Shape = CircleShape): Modifier {
 /** containerColor + containerCornerRadius pair for ToggleFloatingActionButton. */
 data class CelestialFabColors(
   val containerColor: (Float) -> Color,
-  val containerCornerRadius: () -> Dp,
+  val containerCornerRadius: (Float) -> Dp,
 )
 
 @Composable
